Introduce lock schema
