+++ packages/mambajs-core/schema/lock.v1.0.0.json
@@ -1,0 +1,130 @@
+{
+  "title": "ILockV1",
+  "type": "object",
+  "additionalProperties": false,
+  "required": ["lock.version", "platform", "specs", "channels", "packages"],
+  "properties": {
+    "lock.version": {
+      "type": "string",
+      "enum": ["1.0.0"]
+    },
+    "platform": {
+      "type": "string",
+      "enum": [
+        "noarch",
+        "linux-32",
+        "linux-64",
+        "linux-aarch64",
+        "linux-armv6l",
+        "linux-armv7l",
+        "linux-ppc64le",
+        "linux-ppc64",
+        "linux-ppc",
+        "linux-s390x",
+        "linux-riscv32",
+        "linux-riscv64",
+        "osx-64",
+        "osx-arm64",
+        "win-32",
+        "win-64",
+        "win-arm64",
+        "emscripten-wasm32",
+        "wasi-wasm32",
+        "zos-z"
+      ]
+    },
+    "specs": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "channel_priority": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "channels": {
+      "type": "object",
+      "patternProperties": {
+        "^[a-zA-Z0-9._-]+$": {
+          "type": "array",
+          "items": {
+            "type": "object",
+            "required": ["url", "protocol"],
+            "properties": {
+              "url": {
+                "type": "string",
+                "format": "uri"
+              },
+              "protocol": {
+                "type": "string",
+                "enum": ["https"]
+              }
+            },
+            "additionalProperties": false
+          }
+        }
+      },
+      "additionalProperties": false
+    },
+    "packages": {
+      "type": "array",
+      "items": {
+        "type": "object",
+        "patternProperties": {
+          ".*": {
+            "type": "object",
+            "required": ["name", "build_string", "version", "subdir", "channel"],
+            "properties": {
+              "name": {
+                "type": "string"
+              },
+              "build_string": {
+                "type": "string"
+              },
+              "version": {
+                "type": "string"
+              },
+              "subdir": {
+                "type": "string"
+              },
+              "channel": {
+                "type": "string"
+              }
+            },
+            "additionalProperties": false
+          }
+        },
+        "additionalProperties": false
+      }
+    },
+    "pipPackages": {
+      "type": "array",
+      "items": {
+        "type": "object",
+        "patternProperties": {
+          ".*": {
+            "type": "object",
+            "required": ["name", "version", "url"],
+            "properties": {
+              "name": {
+                "type": "string"
+              },
+              "version": {
+                "type": "string"
+              },
+              "url": {
+                "type": "string",
+                "format": "uri"
+              }
+            },
+            "additionalProperties": false
+          }
+        },
+        "additionalProperties": false
+      }
+    }
+  }
+}
