+++ packages/mambajs-core/src/helper.ts
@@ -1,78 +1,6 @@
 import { FilesData, IUnpackJSAPI } from '@emscripten-forge/untarjs';
 import { parse } from 'yaml';
-
-export interface ILogger {
-  log(...msg: any[]): void;
-  warn(...msg: any[]): void;
-  error(...msg: any[]): void;
-}
-
-export interface ISolvedPackage {
-  name: string;
-  version: string;
-  repo_url?: string;
-  url: string;
-  build_number?: number;
-  repo_name?: string;
-  build_string?: string;
-  subdir?: string;
-  depends?: string[];
-}
-
-export interface ISolvedPackages {
-  [key: string]: ISolvedPackage;
-}
-
-export interface IEnvPackages {
-  condaPackages: ISolvedPackages;
-  pipPackages: ISolvedPackages;
-}
-
-export interface IEnv {
-  specs: string[];
-  channels: string[];
-  packages: IEnvPackages;
-}
-
-export interface IEmpackEnvMetaPkg {
-  name: string;
-  version: string;
-  build: string;
-  channel: string;
-  filename_stem: string;
-  filename: string;
-  url: string;
-  depends: [];
-  subdir: string;
-}
-
-export interface IEmpackEnvMetaMountPoint {
-  name: string;
-  filename: string;
-}
-
-export interface IEmpackEnvMeta {
-  prefix: string;
-  packages: IEmpackEnvMetaPkg[];
-  specs?: string[];
-  channels?: string[];
-  mounts?: IEmpackEnvMetaMountPoint[];
-}
-
-/**
- * Shared libraries. list of .so files
- */
-export type TSharedLibs = string[];
-
-/**
- * Shared libraries. A map package name -> list of .so files
- */
-export type TSharedLibsMap = { [pkgName: string]: TSharedLibs };
-export interface IBootstrapData {
-  sharedLibs: TSharedLibsMap;
-  paths: { [key: string]: string };
-  untarjs: IUnpackJSAPI;
-}
+import { ISolvedPackages, TSharedLibs } from './types';
 
 export function parseEnvYml(envYml: string) {
   const data = parse(envYml);
